# docs/openapi/api-gateway.json
{
  "components": {
    "schemas": {
...
      "TrebovanjeDetail": {
        "properties": {
          "broj_stavki": {
            "type": "integer"
          },
          "datum": {
            "format": "date-time",
            "type": "string"
          },
          "dokument_broj": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "magacin": {
            "type": "string"
          },
          "radnja": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "stavke": {
            "items": {
              "$ref": "#/components/schemas/TrebovanjeItemDetail"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TrebovanjeImportPayload": {
        "properties": {
          "datum": {
            "format": "date-time",
            "type": "string"
          },
          "dokument_broj": {
            "type": "string"
          },
          "magacin_naziv": {
            "type": "string"
          },
          "magacin_pantheon_id": {
            "type": "string"
          },
          "radnja_naziv": {
            "type": "string"
          },
          "radnja_pantheon_id": {
            "type": "string"
          },
          "stavke": {
            "items": {
              "properties": {
                "artikl_sifra": {
                  "type": "string"
                },
                "barkod": {
                  "type": "string"
                },
                "kolicina_trazena": {
                  "type": "number"
                },
                "naziv": {
                  "type": "string"
                }
              },
              "required": [
                "artikl_sifra",
                "naziv",
                "kolicina_trazena"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "dokument_broj",
          "datum",
          "magacin_pantheon_id",
          "radnja_pantheon_id",
          "stavke"
        ],
        "type": "object"
      },
      "TrebovanjeItemDetail": {
        "properties": {
          "artikl_sifra": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kolicina_trazena": {
            "type": "number"
          },
          "kolicina_uradjena": {
            "type": "number"
          },
          "naziv": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TrebovanjeListItem": {
        "properties": {
          "broj_stavki": {
            "type": "integer"
          },
          "datum": {
            "format": "date-time",
            "type": "string"
          },
          "dokument_broj": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "magacin": {
            "type": "string"
          },
          "radnja": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "ukupno_trazena": {
            "type": "number"
          },
          "ukupno_uradjena": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "dokument_broj"
        ],
        "type": "object"
      },
      "TrebovanjeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrebovanjeListItem"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ZaduznicaAssignment": {
        "properties": {
          "due_at": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ZaduznicaItemPayload"
            },
            "type": "array"
          },
          "magacioner_id": {
            "format": "uuid",
            "type": "string"
          },
          "priority": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ZaduznicaCreateRequest": {
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/ZaduznicaAssignment"
            },
            "type": "array"
          },
          "trebovanje_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ZaduznicaItemPayload": {
        "properties": {
          "quantity": {
            "type": "number"
          },
          "trebovanje_stavka_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "trebovanje_stavka_id",
          "quantity"
        ],
        "type": "object"
      },
      "ZaduznicaReassignRequest": {
        "properties": {
          "target_magacioner_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "target_magacioner_id"
        ],
        "type": "object"
      },
      "ZaduznicaStatusUpdate": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Magacin API Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/trebovanja": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "default": 20,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "magacin_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "radnja_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrebovanjeListResponse"
                }
              }
            },
            "description": "Paged result"
          }
        },
        "summary": "List trebovanja"
      }
    },
    "/api/trebovanja/import": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrebovanjeImportPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Import trebovanje"
      }
    },
    "/api/trebovanja/{trebovanje_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "trebovanje_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrebovanjeDetail"
                }
              }
            },
            "description": "Trebovanje detail"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Get trebovanje detail"
      }
    },
...
    "/api/zaduznice": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaduznicaCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create zaduznice"
      }
    },
    "/api/zaduznice/{zaduznica_id}/reassign": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaduznicaReassignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated"
          }
        },
        "summary": "Reassign zaduznica"
      }
    },
    "/api/zaduznice/{zaduznica_id}/status": {
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "zaduznica_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaduznicaStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated"
          }
        },
        "summary": "Update zaduznica status"
      }
    }
  }
}